feat(app): set backend demo as home; add /client-demo; update footer
- Swap homepage to backend demo (moved app/backend-demo/page.tsx → app/page.tsx export)
- Add app/client-demo/page.tsx rendering existing client SDK flow
- Remove header link; footer now shows “Client SDK demo” and links to /client-demo
- Footer “View code” updated to my repo; removed Crossmint quickstart/crossmint.com links
- Verified locally: / backend UI; /client-demo client UI; API calls target NEXT_PUBLIC_BACKEND_URL
