+++ components/footer.tsx
@@ -6,7 +6,7 @@ export function Footer() {
       <div className="flex gap-6 flex-wrap items-center justify-center">
         <a
           className="flex items-center gap-2 hover:underline hover:underline-offset-4"
-          href="https://github.com/Crossmint/wallets-quickstart"
+          href="https://github.com/TomCat-415/wallets-quickstart"
           target="_blank"
           rel="noopener noreferrer"
         >
@@ -21,7 +21,7 @@ export function Footer() {
         </a>
         <a
           className="flex items-center gap-2 hover:underline hover:underline-offset-4"
-          href="/backend-demo"
+          href="/client-demo"
         >
           <Image
             aria-hidden
@@ -30,38 +30,9 @@ export function Footer() {
             width={16}
             height={16}
           />
-          Backend Demo (Server API)
+          Client SDK demo
         </a>
-        <a
-          className="flex items-center gap-2 hover:underline hover:underline-offset-4"
-          href="https://www.crossmint.com/quickstarts"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          <Image
-            aria-hidden
-            src="/window.svg"
-            alt="Window icon"
-            width={16}
-            height={16}
-          />
-          See all quickstarts
-        </a>
-        <a
-          className="flex items-center gap-2 hover:underline hover:underline-offset-4"
-          href="https://crossmint.com"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          <Image
-            aria-hidden
-            src="/globe.svg"
-            alt="Globe icon"
-            width={16}
-            height={16}
-          />
-          Go to crossmint.com →
-        </a>
+        
       </div>
       <div className="flex">
         <Image
